moved canvasjs-files to app/javascript/lib
