+++ app/javascript/lib/canvasjs/canvasjs.react.js
